+++ src/main/java/com/swp391/edrive/service/VehicleService.java
@@ -5,6 +5,7 @@ import com.swp391.edrive.dto.response.VehicleResponse;
 import com.swp391.edrive.entity.Vehicle;
 import com.swp391.edrive.enums.VehicleStatus;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface VehicleService {
@@ -21,7 +22,7 @@ public interface VehicleService {
 
     List<VehicleResponse> findVehicleByManufactureYearRange(Integer fromYear, Integer toYear, int page, int size);
 
-    List<VehicleResponse> findVehicleByPrice(Double minPrice, Double maxPrice, int page, int size);
+    List<VehicleResponse> findVehicleByPrice(BigDecimal minPrice, BigDecimal maxPrice, int page, int size);
 
     VehicleResponse createVehicle(VehicleUpsertRequest req);
